fix: close dependency, security, testing, and ci gaps
- close dependency, security, testing, and CI gaps
- resolve 16 issues across security, correctness, and lint

+++ crates/crypto-engine/src/threshold/refresh/protocol.rs
@@ -506,7 +506,7 @@ impl KeyRefreshProtocol {
                         &mut result,
                         generator,
                         coeff_bytes.as_ptr(),
-                        (coeff_bytes.len() * 8) as usize,
+                        coeff_bytes.len() * 8,
                     );
 
                     // Compress the result
